requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==4.1.0
black==22.6.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-timezone-field==5.1
djangorestframework==3.14.0
isort==5.12.0
kombu==5.3.1
mypy-extensions==1.0.0
pathspec==0.11.1
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
